chore: rename JavaScript files to follow kebab-case
